requirements.txt
discord.py>=2.0.0
requests>=2.25.1
Pillow>=8.2.0
numpy>=1.20.0
python-dotenv>=0.19.0